Setup AdminBundle as a Node.js package
