+++ src/Sylius/Bundle/AdminBundle/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@sylius-ui/admin",
+  "version": "1.0.0",
+  "license": "MIT",
+  "author": "Sylius Sp. z o.o.",
+  "dependencies": {
+    "@popperjs/core": "^2.11.8",
+    "@symfony/webpack-encore": "^3.1.0",
+    "@tabler/core": "^1.0.0-beta20",
+    "apexcharts": "^3.41.0",
+    "bootstrap": "^5.3.0",
+    "choices.js": "^10.2.0"
+  },
+  "devDependencies": {
+    "sass": "^1.54.8",
+    "sass-loader": "^13.0.0"
+  },
+  "engines": {
+    "node": "^18 || ^20"
+  },
+  "engineStrict": true,
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Sylius/SyliusAdminBundle.git"
+  }
+}
